feat: GetItem on User
This introduces the `GetEntity` and `FormatKey` traits.

The `FormatKey` allows the implementor to specify a type, `Key` that will be
used as the object's key, and includes the `format_key` which is called
with a key in order to receive the attributes needed by the database.
FormatKey definitions are included in the proc macros alongside `GetKeys`.

The `GetEntity` trait receives a key's attribute map and prepares
a `GetItem` query that will retrieve the item with that key.