feat: Added Application entity type
This is an example of an `OwnedEntity`. It is namespaced under an
key that is to be passed in when querying.

In order to make querying work, the query trait had to be changed to
accept the namespace to query under. For `UniqueEntity`, there currently
is no namespace (so the value passed is `()`).

Instead of implementing the query over a generic T it was moved into the
derive proc macro and is implemented for each class. I think this is
working cleaner than trying to make a more simple trait from which
querying is derived from. It may be a good idea to consider deriving the other
traits in a similar manner.