feat: token generation
The authentication and signup routes now return a real token.
In order to facilitate this, the server now requires the public and
private key to be specified. Instead of expanding the `new` function,
the the `Server` struct now has a `builder` function which intiiates a
builder to specify what it needs.

The server now exposes `/.well-known/jwks.json` which will contain the
public key.

In order to store keys, the axum State is now a struct containing the
keys (and the previous state which was the dynamodb table info).