feat: Load development environment variables
Direnv is leveraged to load environment variables into development
shells. `bin/direnvrc` has been created as a location to place direnv
extension functionality. A script has been placed there that will load
all of the files contained in `.direnv/env` as environment variables
with the file name used as the variable name and the file contents as
values. The `.direnv` directory is included in `.gitignore`, so these
variables will not be checked in to version control.