feat: load configuration from environment
The `config` crate has been included to help load starting values
(currently database connection information and JWT keys). Values will be
loaded from the environment with the prefix "LOCKPAD_". The JWT keys are
no longer loaded in from a file, but instead their entire contents are
stored in the environment variable passed in.