fix: Make tests ignore token expiration date
The tests that decode a JWT have a hardcoded token they use. This worked
fine initially, but after the time of the token's expiration the test
failed. For now the easiest solution is to skip validating that field.
To facilitate that, the `Claims` struct now has a `decode_validation`
function that operates like the `decode` function but allows for passing
in a custom `jsonwebtoken::Validation` that will be used instead of
creating a default one.