feat: add services-flake for process-compose
This adds a new package to the project that will run process-compose
when executed. The process-compose configuration is set up using
`services-flake`. Currently the only process is postgres which is
intended to replace the self-rolled `start-database` script that the
project uses for postgres. Ideally this will allow for a similar
use-case.

The primary difference between this process-compose strategy and the
existing one is that the process-compose version uses a networking port
and thus can conflict if running multiple projects this way. The
existing implementation uses a unix socket for connecting to the
database and thus allows for multiple projects to be running databases
this method at once without negotiating for port numbers.